refactor: Improve fit
